Splitting index.ts up and added basic Proplists code
